[kernel] Build everything with -Wconversion [part 7]
Changes to be committed:
	modified:   include/tilck/common/arch/generic_x86/cpu_features.h
	modified:   kernel/arch/generic_x86/cpu.c
	modified:   kernel/arch/generic_x86/mtrr.c
	modified:   kernel/arch/i386/paging.c

+++ kernel/arch/i386/paging.c
@@ -818,5 +818,5 @@ void map_framebuffer(uptr paddr, uptr vaddr, uptr size, bool user_mmap)
       return;
    }
 
-   set_mtrr(selected_mtrr, paddr, pow2size, MEM_TYPE_WC);
+   set_mtrr((u32)selected_mtrr, paddr, pow2size, MEM_TYPE_WC);
 }
